backend/requirements.txt
django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg[binary]==3.2.3
dj-database-url==2.1.0
gunicorn==21.2.0
python-dotenv==1.0.1
boto3==1.34.69
django-storages==1.14.3
